Implement reset logic in connector host as 'reset' method here
Summary: The connector host hard-resets DDP's underlying websocket connection by playing with the libraries internal properties (this is why it broke when platform upgraded to 0.12.1). This accomplishes the same thing, but implemented as a method in the library itself, since we're maintaining a fork to get proxy support anyhow.

Test Plan: Tested manually with dev connector host by mocking missed heartbeat acks. I tested the version platform was using (0.9.1), the version they had issues with (0.12.1), the original fork (newer than 0.12.1), and this version. All work (no discernable behavior difference) except 0.12.1, which fails in a way that's consistent with platform team's observations.

Reviewers: micahrye

Reviewed By: micahrye

Differential Revision: https://phab.tulip.co/D3204